Rename package `set` to `models`
