+++ app/src/main/java/crux/bphc/cms/fragments/ForumFragment.java
@@ -37,7 +37,7 @@ import io.realm.Realm;
 import io.realm.RealmResults;
 import retrofit2.Retrofit;
 import retrofit2.converter.gson.GsonConverterFactory;
-import set.forum.Discussion;
+import models.forum.Discussion;
 
 import static crux.bphc.cms.app.Constants.API_URL;
