+++ app/src/main/java/models/Content.java
@@ -1,4 +1,4 @@
-package set;
+package models;
 
 import io.realm.RealmObject;
 import io.realm.annotations.Index;
