+++ app/src/main/java/models/CourseSection.java
@@ -1,4 +1,4 @@
-package set;
+package models;
 
 import android.text.Html;
