+++ app/src/main/java/models/Module.java
@@ -1,4 +1,4 @@
-package set;
+package models;
 
 import android.text.Html;
