+++ app/src/main/java/models/enrol/SelfEnrolWarning.java
@@ -1,4 +1,4 @@
-package set.enrol;
+package models.enrol;
 
 /**
  * Created by siddhant on 12/19/16.
