+++ app/src/main/java/models/forum/Attachment.java
@@ -1,4 +1,4 @@
-package set.forum;
+package models.forum;
 
 import io.realm.RealmObject;
