+++ app/src/main/java/models/forum/Discussion.java
@@ -1,4 +1,4 @@
-package set.forum;
+package models.forum;
 
 import android.text.Html;
